Fix ENV load
A little bit of a miscommunication between the README and
implementation...the bin/setup creates an `.env.development` file, but
the code was not loading this while also erroring if `.env` didn't
exist.

The config package will now load `.env.<environment>` if it exists,
while also not erroring if `.env` doesn't exist.

The ENV variable `APP_ENV` has also been introduced, with the purpose
for now of determining which `.env` file to load.